Fix get licenses in extension
