+++ src/extension/src/client-modules/generation.ts
@@ -3,15 +3,9 @@ import { WizardServant, IPayloadResponse } from "../wizardServant";
 import { CONSTANTS } from "../constants/constants";
 import { TelemetryEventName } from "../constants/telemetry";
 import { ITelemetryService } from "../telemetry/telemetryService";
-import {
-  IAppService,
-  ICosmosDB,
-  IGenerationData,
-  SERVICE_CATEGORY,
-  SERVICE_TYPE,
-} from "../types/generationTypes";
 import { EXTENSION_COMMANDS } from "../constants/commands";
 import GenerationService from "../utils/generation/GenerationService";
+import { getGenerationData } from "../utils/generation/generationUtils";
 
 export class Generation extends WizardServant {
   clientCommandMap: Map<EXTENSION_COMMANDS, (message: any) => Promise<IPayloadResponse>> = new Map([
@@ -26,7 +20,7 @@ export class Generation extends WizardServant {
   private async generate(message: any): Promise<IPayloadResponse> {
     this.trackWizardTotalSessionTimeToGenerate();
     const generationService = new GenerationService(this.Telemetry);
-    const generationData = this.getGenerationData(message.payload);
+    const generationData = getGenerationData(message.payload);
     await generationService.generate(generationData);
     return { payload: undefined };
   }
@@ -46,51 +40,5 @@ export class Generation extends WizardServant {
       this.Telemetry.wizardSessionStartTime,
       Date.now()
     );
-  }
-
-  private getGenerationData(data: any): IGenerationData {
-    const generationData: IGenerationData = {
-      backendFramework: data.backendFramework,
-      frontendFramework: data.frontendFramework,
-      backendFrameworkLinuxVersion: data.backendFrameworkLinuxVersion,
-      pages: data.pages,
-      path: data.path,
-      projectName: data.projectName,
-      projectType: data.projectType,
-      services: [],
-    };
-
-    if (data.services?.appService) {
-      const { internalName, subscription, resourceGroup, location, siteName } = data.services.appService;
-
-      const appService: IAppService = {
-        internalName,
-        category: SERVICE_CATEGORY.AZURE,
-        type: SERVICE_TYPE.APPSERVICE,
-        subscription,
-        resourceGroup,
-        location,
-        serviceName: siteName,
-      };
-      generationData.services?.push(appService);
-    }
-
-    if (data.services?.cosmosDB) {
-      const { internalName, subscription, resourceGroup, location, accountName, api } = data.services.cosmosDB;
-
-      const cosmosDB: ICosmosDB = {
-        internalName,
-        category: SERVICE_CATEGORY.AZURE,
-        type: SERVICE_TYPE.COSMOSDB,
-        subscription,
-        resourceGroup,
-        location,
-        serviceName: accountName,
-        api,
-      };
-      generationData.services?.push(cosmosDB);
-    }
-
-    return generationData;
   }
 }
